Develop

+++ PBL3_NetManagement/GUI/FormAdmin.cs
@@ -19,12 +19,14 @@ namespace PBL3_NetManagement
         public del1 SetVisible_Login;
         DateTime login_time;
         Timer timer_LoadComputer;
+        Bill currentBill;
         public FormAdmin(string username)
         {
             InitializeComponent();
             login_time = DateTime.Now;
             AdminName = username;
             BLL_NM.Instance.Login_init(login_time, AdminName);
+            currentBill = BLL_NM.Instance.Get_Newest_Bill();
         }
         private void Logout()
         {
@@ -45,12 +47,19 @@ namespace PBL3_NetManagement
 
         private void buttonDeleteComputer_Click(object sender, EventArgs e)
         {
+            if (string.Equals(textBoxIDComputer.Text, "")) return;
             if ((buttonDeleteComputer.Tag as Computer).ComputerStatus == true)
             {
                 MessageBox.Show("This computer is currently in use!");
                 return;
             }
-            BLL_NM.Instance.Delete_Computer((buttonDeleteComputer.Tag as Computer).idComputer);
+            DialogResult result = MessageBox.Show("Do you really want to delete this computer", "Warning", MessageBoxButtons.YesNo);
+            if(result == System.Windows.Forms.DialogResult.Yes)
+            {
+                BLL_NM.Instance.Delete_Computer_Log((buttonDeleteComputer.Tag as Computer).idComputer);
+                BLL_NM.Instance.Delete_Computer((buttonDeleteComputer.Tag as Computer).idComputer);
+                LoadSystemLogs();
+            }
             this.textBoxNameComputer.Text = "";
             this.textBoxStatusComputer.Text = "";
             this.textBoxPriceComputer.Text = "";
@@ -59,6 +68,7 @@ namespace PBL3_NetManagement
 
         private void buttonEditComputer_Click(object sender, EventArgs e)
         {
+            if (string.Equals(textBoxIDComputer.Text, "")) return;
             FormAddEditComputer faec = new FormAddEditComputer(buttonEditComputer.Tag as Computer);
             faec.Show();
         }
@@ -67,10 +77,25 @@ namespace PBL3_NetManagement
         {
             timer_LoadComputer = new Timer();
             timer_LoadComputer.Tick += LoadComputer_Tick;
+            timer_LoadComputer.Tick += Check_newBill;
             timer_LoadComputer.Interval = 1000;
             timer_LoadComputer.Enabled = true;
             Load_Table();
             LoadSystemLogs();
+            LoadBill();
+            LoadAccounts();
+            Load_DTGrid_Good();
+        }
+        // Kiểm tra xem có Bill mới hay không
+        private void Check_newBill(object sender, EventArgs e)
+        {
+            Bill nextBill = BLL_NM.Instance.Get_Newest_Bill();
+            if (nextBill.idBill > currentBill.idBill)
+            {
+                currentBill.idBill += 1;
+                MessageBox.Show("idBill: " + nextBill.idBill + "\nUsername: " + nextBill.UserName, "New order!");
+            }
+            else return;
         }
         // Khi timer load lại flowl thì đồng thời load lại các textbox chứa thông tin của máy
         private void LoadComputer_Tick(object sender, EventArgs e)
@@ -89,8 +114,8 @@ namespace PBL3_NetManagement
             {
                 Button bt = new Button() { Width = 75, Height = 75 };
                 string status;
-                if (item.ComputerStatus == false) status = "Trống";
-                else status = "Bận";
+                if (item.ComputerStatus == false) status = "OFF";
+                else status = "ON";
                 bt.Text = item.ComputerName + Environment.NewLine + status;
                 bt.Click += Bt_Click;
                 bt.Tag = item;
@@ -108,9 +133,7 @@ namespace PBL3_NetManagement
         // load thông tin Computer lên các textbox
         private void Load_Info_Computer(string idcomputer)
         {
-            // fix exception
             if (!BLL_NM.Instance.ComputerCheck(idcomputer)) return;
-
             Computer computer = BLL_NM.Instance.Get_Computer(idcomputer);
             textBoxIDComputer.Text = computer.idComputer.ToString();
             textBoxNameComputer.Text = computer.ComputerName.ToString();
@@ -138,7 +161,7 @@ namespace PBL3_NetManagement
             foreach(ComputerLog i in BLL_NM.Instance.GetComputerLogs().OrderByDescending(o => o.idLog))
             {
                 if (!i.UserName.Contains(textBox_SystemLogSearch.Text)) continue;
-                textBoxSystemLog_All.Text += "Log Id: " + i.idLog + "\r\n";
+                textBoxSystemLog_All.Text += "Log Id:     " + i.idLog + "\r\n";
                 textBoxSystemLog_All.Text += "Username:     " + i.UserName + "\r\n";
                 textBoxSystemLog_All.Text += "Computer Id:  " + i.idComputer + "\r\n";
                 textBoxSystemLog_All.Text += "Login time:     " + i.DateLogin + "\r\n";
@@ -148,7 +171,12 @@ namespace PBL3_NetManagement
                 textBoxSystemLog_All.Text += "-----------------------------------------------------\r\n";
             }
         }
-
+        private void Load_DTGrid_Good()
+        {
+            dataGridView_Goods.Controls.Clear();
+            dataGridView_Goods.DataSource = BLL_NM.Instance.Get_All_Good_With_Name(textBoxSearch_Goods.Text);
+            dataGridView_Goods.Columns["idGood"].Visible = false;
+        }
         private void buttonRefreshSystemLog_Click(object sender, EventArgs e)
         {
             LoadSystemLogs();
@@ -157,6 +185,126 @@ namespace PBL3_NetManagement
         private void textBox_SystemLogSearch_TextChanged(object sender, EventArgs e)
         {
             LoadSystemLogs();
+        }
+        private void LoadBill()
+        {
+            textBoxBill.Text = "";
+            foreach (Bill lbill in BLL_NM.Instance.Get_Bill().OrderByDescending(o => o.idBill))
+            {
+                if (!lbill.UserName.Contains(texboxSearch_Bills.Text)) continue;
+                textBoxBill.Text += "Bill ID: " + lbill.idBill + "\r\n";
+                textBoxBill.Text += "Username: " + lbill.UserName + "\r\n";
+                textBoxBill.Text += "Time: " + lbill.Date + "\r\n";
+                foreach (BillInfo lbillinfo in BLL_NM.Instance.Get_Billinfo_with_idBill(lbill.idBill))
+                {
+                    textBoxBill.Text += "ID BillIfo: " + lbillinfo.idBillInfo + "    ";
+                    textBoxBill.Text += "ID Good: " + lbillinfo.idGood + "    ";
+                    textBoxBill.Text += "Good Name: " + BLL_NM.Instance.Text_alignment(BLL_NM.Instance.Get_GoodName(Convert.ToInt32(lbillinfo.idGood.ToString())),20);
+                    textBoxBill.Text += "Count: " + lbillinfo.Count + "\r\n";
+                }
+                textBoxBill.Text += "-----------------------------------------------------\r\n";
+            }
+        }
+
+        private void textboxBill_Textchanged(object sender, EventArgs e)
+        {
+            LoadBill();
+        }
+
+        private void buttonRefresh_Click(object sender, EventArgs e)
+        {
+            LoadBill();
+        }
+        private void LoadAccounts()
+        {
+            dataGridView_Account.DataSource = BLL_NM.Instance.Get_Clients_Show(textBoxSearch_Account.Text);
+        }
+
+        private void textBoxSearch_Account_TextChanged(object sender, EventArgs e)
+        {
+            LoadAccounts();
+        }
+
+        private void buttonDep_Account_Click(object sender, EventArgs e)
+        {
+            if (dataGridView_Account.CurrentRow == null) return;
+            FormDep fd = new FormDep(dataGridView_Account.CurrentRow.Cells["Username"].Value.ToString());
+            fd.Reload_Accounts = LoadAccounts;
+            fd.Show();
+        }
+
+        private void buttonAdd_Account_Click(object sender, EventArgs e)
+        {
+            FormAddEditAccount ftemp = new FormAddEditAccount();
+            ftemp.Reload_Accounts = LoadAccounts;
+            ftemp.Show();
+        }
+
+        private void buttonEdit_Account_Click(object sender, EventArgs e)
+        {
+            if (dataGridView_Account.CurrentRow == null) return;
+            FormAddEditAccount ftemp = new FormAddEditAccount(dataGridView_Account.CurrentRow.Cells["Username"].Value.ToString());
+            ftemp.Reload_Accounts = LoadAccounts;
+            ftemp.Show();
+        }
+
+        private void buttonDel_Account_Click(object sender, EventArgs e)
+        {
+            if (dataGridView_Account.CurrentRow == null) return;
+            if (BLL_NM.Instance.GetAccountStatus(dataGridView_Account.CurrentRow.Cells["Username"].Value.ToString()))
+            {
+                MessageBox.Show("This user is using a computer, can't delete!");
+            }
+            DialogResult result = MessageBox.Show("Do you really want to delete this account", "Warning", MessageBoxButtons.YesNo);
+            if (result == System.Windows.Forms.DialogResult.Yes)
+            {
+                BLL_NM.Instance.Delete_Account(dataGridView_Account.CurrentRow.Cells["Username"].Value.ToString());
+                LoadAccounts();
+                LoadBill();
+                LoadSystemLogs();
+            }
+        }
+
+        private void textBoxSearch_Goods_TextChanged(object sender, EventArgs e)
+        {
+            Load_DTGrid_Good();
+        }
+
+        private void buttonAdd_Goods_Click(object sender, EventArgs e)
+        {
+            FormAddEditGood fgood = new FormAddEditGood();
+            fgood.Reload_Goods = Load_DTGrid_Good;
+            fgood.Text = "Add good";
+            fgood.Show();
+        }
+
+        private void buttonEdit_Goods_Click(object sender, EventArgs e)
+        {
+            DataGridViewRow dr = dataGridView_Goods.CurrentRow;
+            if (dr == null) return;
+            FormAddEditGood fgood = new FormAddEditGood(Convert.ToInt32(dr.Cells["idGood"].Value.ToString()));
+            fgood.Reload_Goods = Load_DTGrid_Good;
+            fgood.Text = "Edit good";
+            fgood.Show();
+        }
+
+        private void buttonDel_Goods_Click(object sender, EventArgs e)
+        {
+            DataGridViewRow dr = dataGridView_Goods.CurrentRow;
+            if (dr == null) return;
+            DialogResult result = MessageBox.Show("Do you really want to delete this Good", "Warning", MessageBoxButtons.YesNo);
+            if (result == System.Windows.Forms.DialogResult.Yes)
+            {
+                BLL_NM.Instance.Delete_Good(Convert.ToInt32(dr.Cells["idGood"].Value.ToString()));
+                Load_DTGrid_Good();
+                LoadBill();
+            }
+        }
+
+        private void buttonAddBill_Click(object sender, EventArgs e)
+        {
+            FormOrder od = new FormOrder(dataGridView_Account.CurrentRow.Cells["Username"].Value.ToString());
+            od.Show();
         }
     }
 }
